fix: 숫자 리스트 및 결과값의 자료형 수정
1. int, long 타입 사용시 자료형의 크기를 넘어가는 크기의 수를 처리할 수 없음
2. 기본 자료형 대신 BigInteger 객체로 처리

+++ src/main/java/calculator/Calculator.java
@@ -2,6 +2,7 @@ package calculator;
 
 import camp.nextstep.edu.missionutils.Console;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -26,19 +27,19 @@ public class Calculator {
     }
 
     // 문자열 분리 및 숫자 리스트 반환
-    private List<Integer> getNumList(String input) {
+    private List<BigInteger> getNumList(String input) {
         return separator.separateString(input);
     }
 
     // 합산 결과 출력
-    private void printResult(int result) {
+    private void printResult(BigInteger result) {
         System.out.println("결과 : " + result);
     }
 
     public void run() {
         System.out.println("덧셈할 문자열을 입력해 주세요.");
         String input = getString();
-        List<Integer> numList = getNumList(input);
+        List<BigInteger> numList = getNumList(input);
         printResult(
             new Adder(numList).addAll()
         );
